ngs-analysis-app/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pandas==2.1.4
numpy>=1.26.0
pyyaml==6.0.1
python-multipart==0.0.6
aiofiles==23.2.1
openai==1.12.0
scipy>=1.11.0
python-dotenv==1.0.0
# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
# Database & Caching
sqlalchemy==2.0.23
alembic==1.13.1
redis==5.0.1
# Logging & Monitoring
structlog==23.2.0
# Rate Limiting
slowapi==0.1.9
# Vector Search for RAG
chromadb==0.4.18
sentence-transformers==2.2.2
# PII Detection
presidio-analyzer==2.2.33
presidio-anonymizer==2.2.33
# Additional utilities
httpx==0.25.2
asyncpg==0.29.0